+++ crates/script/src/resume.rs
@@ -1,7 +1,8 @@
+use crate::{broadcast::BundledState, simulate::PreSimulationState};
+
 use super::{
     multi_sequence::MultiChainSequence,
     sequence::{ScriptSequence, ScriptSequenceKind},
-    states::{BundledState, PreSimulationState},
 };
 use ethers_providers::Middleware;
 use eyre::Result;
